✨Refactored error handling utils, renamed files/imports
