+++ ai_commit_msg/services/openai_service.py
@@ -1,5 +1,5 @@
 from logging import Logger
-from ai_commit_msg.utils.error_handling import map_error
+from ai_commit_msg.utils.error import map_error
 from openai import OpenAI
 from ai_commit_msg.services.config_service import ConfigService
 from ai_commit_msg.services.local_db_service import LocalDbService, CONFIG_COLLECTION_KEY
